Fix compilation warnings (#2)
* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* State 2023-03-09 20:08:45

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix compilation warnings

* Fix sector highlight

* Fix saved game load

+++ ja2lib/Editor/EditScreen.c
@@ -505,7 +505,7 @@ BOOLEAN EditModeShutdown(void) {
     gfNeedToInitGame = FALSE;
   } else {
     if (!gfBasement && !gfCaves) LightSetBaseLevel((UINT8)gusGameLightLevel);
-    UpdateMercsInSector(gWorldSectorX, gWorldSectorY, gbWorldSectorZ);
+    UpdateMercsInSector((u8)gWorldSectorX, (u8)gWorldSectorY, gbWorldSectorZ);
   }
 
   gGameOptions.fGunNut = gfRealGunNut;
