+++ ja2lib/Laptop/History.c
@@ -131,8 +131,7 @@ void PerformCheckOnHistoryRecord(UINT32 uiErrorCode, u8 sSectorX, u8 sSectorY, I
 void BtnHistoryDisplayNextPageCallBack(GUI_BUTTON *btn, INT32 reason);
 void BtnHistoryDisplayPrevPageCallBack(GUI_BUTTON *btn, INT32 reason);
 
-UINT32 SetHistoryFact(UINT8 ubCode, UINT8 ubSecondCode, UINT32 uiDate, INT16 sSectorX,
-                      INT16 sSectorY) {
+UINT32 SetHistoryFact(UINT8 ubCode, UINT8 ubSecondCode, UINT32 uiDate, u8 sSectorX, u8 sSectorY) {
   // adds History item to player's log(History List), returns unique id number of it
   // outside of the History system(the code in this .c file), this is the only function you'll ever
   // need
@@ -171,8 +170,8 @@ UINT32 SetHistoryFact(UINT8 ubCode, UINT8 ubSecondCode, UINT32 uiDate, INT16 sSe
   return uiId;
 }
 
-UINT32 AddHistoryToPlayersLog(UINT8 ubCode, UINT8 ubSecondCode, UINT32 uiDate, INT16 sSectorX,
-                              INT16 sSectorY) {
+UINT32 AddHistoryToPlayersLog(UINT8 ubCode, UINT8 ubSecondCode, UINT32 uiDate, u8 sSectorX,
+                              u8 sSectorY) {
   // adds History item to player's log(History List), returns unique id number of it
   // outside of the History system(the code in this .c file), this is the only function you'll ever
   // need
@@ -549,7 +548,7 @@ void OpenAndReadHistoryFile(void) {
   HWFILE hFileHandle;
   UINT8 ubCode, ubSecondCode;
   UINT32 uiDate;
-  INT16 sSectorX, sSectorY;
+  i16 sSectorX, sSectorY;
   INT8 bSectorZ = 0;
   UINT8 ubColor;
   UINT32 iBytesRead = 0;
@@ -588,12 +587,12 @@ void OpenAndReadHistoryFile(void) {
 
 #ifdef JA2TESTVERSION
     // perform a check on the data to see if it is pooched
-    PerformCheckOnHistoryRecord(1, sSectorX, sSectorY, bSectorZ);
+    PerformCheckOnHistoryRecord(1, (u8)sSectorX, (u8)sSectorY, bSectorZ);
 #endif
 
     // add transaction
-    ProcessAndEnterAHistoryRecord(ubCode, uiDate, ubSecondCode, sSectorX, sSectorY, bSectorZ,
-                                  ubColor);
+    ProcessAndEnterAHistoryRecord(ubCode, uiDate, ubSecondCode, (u8)sSectorX, (u8)sSectorY,
+                                  bSectorZ, ubColor);
 
     // increment byte counter
     uiByteCount += SIZE_OF_HISTORY_FILE_RECORD;
@@ -622,7 +621,7 @@ BOOLEAN OpenAndWriteHistoryFile(void) {
   while (pHistoryList) {
 #ifdef JA2TESTVERSION
     // perform a check on the data to see if it is pooched
-    PerformCheckOnHistoryRecord(2, pHistoryList->sSectorX, pHistoryList->sSectorY,
+    PerformCheckOnHistoryRecord(2, (u8)pHistoryList->sSectorX, (u8)pHistoryList->sSectorY,
                                 pHistoryList->bSectorZ);
 #endif
 
@@ -766,7 +765,7 @@ void DrawHistoryRecordsText(void) {
                                 pHistoryLocations[0], HISTORY_TEXT_FONT, &sX, &sY);
       mprintf(sX, RECORD_Y + (iCounter * (BOX_HEIGHT)) + 3, pHistoryLocations[0]);
     } else {
-      GetSectorIDString(pCurHistory->sSectorX, pCurHistory->sSectorY, pCurHistory->bSectorZ,
+      GetSectorIDString((u8)pCurHistory->sSectorX, (u8)pCurHistory->sSectorY, pCurHistory->bSectorZ,
                         sString, ARR_SIZE(sString), TRUE);
       FindFontCenterCoordinates(RECORD_DATE_X + RECORD_DATE_WIDTH, 0, RECORD_LOCATION_WIDTH + 10, 0,
                                 sString, HISTORY_TEXT_FONT, &sX, &sY);
@@ -1107,7 +1106,7 @@ BOOLEAN LoadInHistoryRecords(UINT32 uiPage) {
   INT32 iCount = 0;
   HWFILE hFileHandle;
   UINT8 ubCode, ubSecondCode;
-  INT16 sSectorX, sSectorY;
+  i16 sSectorX, sSectorY;
   INT8 bSectorZ;
   UINT32 uiDate;
   UINT8 ubColor;
@@ -1161,12 +1160,12 @@ BOOLEAN LoadInHistoryRecords(UINT32 uiPage) {
 
 #ifdef JA2TESTVERSION
     // perform a check on the data to see if it is pooched
-    PerformCheckOnHistoryRecord(3, sSectorX, sSectorY, bSectorZ);
+    PerformCheckOnHistoryRecord(3, (u8)sSectorX, (u8)sSectorY, bSectorZ);
 #endif
 
     // add transaction
-    ProcessAndEnterAHistoryRecord(ubCode, uiDate, ubSecondCode, sSectorX, sSectorY, bSectorZ,
-                                  ubColor);
+    ProcessAndEnterAHistoryRecord(ubCode, uiDate, ubSecondCode, (u8)sSectorX, (u8)sSectorY,
+                                  bSectorZ, ubColor);
 
     // increment byte counter
     uiByteCount += SIZE_OF_HISTORY_FILE_RECORD;
@@ -1247,7 +1246,7 @@ BOOLEAN WriteOutHistoryRecords(UINT32 uiPage) {
   while ((iCount < NUM_RECORDS_PER_PAGE) && (fOkToContinue)) {
 #ifdef JA2TESTVERSION
     // perform a check on the data to see if it is pooched
-    PerformCheckOnHistoryRecord(4, pList->sSectorX, pList->sSectorY, pList->bSectorZ);
+    PerformCheckOnHistoryRecord(4, (u8)pList->sSectorX, (u8)pList->sSectorY, pList->bSectorZ);
 #endif
 
     FileMan_Write(hFileHandle, &(pList->ubCode), sizeof(UINT8), NULL);
@@ -1397,7 +1396,7 @@ BOOLEAN AppendHistoryToEndOfFile(HistoryUnitPtr pHistory) {
 
 #ifdef JA2TESTVERSION
   // perform a check on the data to see if it is pooched
-  PerformCheckOnHistoryRecord(5, pHistoryList->sSectorX, pHistoryList->sSectorY,
+  PerformCheckOnHistoryRecord(5, (u8)pHistoryList->sSectorX, (u8)pHistoryList->sSectorY,
                               pHistoryList->bSectorZ);
 #endif
